Confirm by Sheldon

* commit '5afb24665c75f7d984660e06e5b3a3a8579b1d0e':
  DHDSMA-8 | Completed Edit

+++ dailyHOTEL/app/src/main/java/com/twoheart/dailyhotel/adapter/BookingListAdapter.java
@@ -128,7 +128,7 @@ public class BookingListAdapter extends ArrayAdapter<Booking> implements PinnedS
         {
             Integer tag = (Integer) convertView.getTag();
 
-            if (tag == null || tag != booking.type)
+            if (tag == null || tag.intValue() != booking.type)
             {
                 convertView = null;
             }
